created Dashboard Directory inside of Pages, and moved Dashboard component, updated imports in apps component.
